pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "Routed multi-specialist AI system for High Energy Physics research"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
serving = [
    "vllm>=0.8.0",
    "transformers>=4.48.0",
    "peft>=0.14.0",
    "torch>=2.5.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.0.0",
]
training = [
    "torch>=2.5.0",
    "transformers>=4.48.0",
    "deepspeed>=0.16.0",
    "datasets>=3.0.0",
    "wandb>=0.19.0",
    "peft>=0.14.0",
    "accelerate>=1.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
